Implement the is_file_updated internal function
This function checks that the submitted filename was modified wrt
previous mtime and size data saved in a "stamp" file located in the
current execution directory. If either the stamp file did not exist or
if its content is different of the data returned by the "stat" system
function, then the file is considered as updated.