Change where the temporary data are stored
The "time stamp" of the input file are stored in the ".htrdr" directory
located in the current execution directory. The cached SVX grids are
stored in the ".htrdr/<MIE>/<GOP>/" directory whith <MIE> and <GOP>
the name of the file storing the Mie's data and the Gas Optical
properties used to compute the aforementioned SVX grids.