Fix the update of the file stamps
The stamp of a given file was updated on the invocation of the
"is_file_updated" function. This function do not update the file stamp
anymore. One have to explicitly call the "update_file_stamp" function to
update the stamp of a given file.

The stamps of the HTCP, HTGOP and HTMIE files are now updated *after*
the generation of the "cloud grids" in order to avoid unforeseen
behaviors. If the htrdr process was killed during the generation
of the grids, the next invocation of the htrdr program might skip the
generation of some grids built from outdated HTCP/HTGOP/HTMIE
files.