Fix htrdr_compute_radiance_sw
Attach filter function to the ground geometry to avoid "self hits".

Handle the case where the ground is not a plane below the sky. In this
case, the random walk can reach a surface that does not front face the
sun. Furthermore, the sun can be occluded by a surface at any scattering
position.