By default, do not use the "cached grids"
Previously, the program computes grids of optical properties of the
clouds for each spectral band and for each quadrature point. Once built,
the grids are stored onto disk and, as far as it is possible, are
reused between runs.

Each cached grid can take up a lot of space. As a consequence, depending
on the disk efficiency and the available system memory, the loading of
the cached data can be time consuming, annihilating the benefit of the
caching mechanism.

By default, the program now directly makes parallel the building of the
cloud properties octrees, not the generation of the intermediary cached
grids. In some situation, this strategy speeds up the setup of clouds up
to a factor of 2, even in situation where the grids were already
precomputed.

To still rely on the grid caching mechanism, use the new '-G' option.