Improve work stealing strategy
Previously, the draw function was waiting for the completion of the
rendering threads and then submitted a steal request to the concurrent
processes. Some rendering threads were thus waiting for the completion
of the other rendering threads without performing any computation.

In this commit a thread with no more work does not wait anymore and
directly try to steal tiles to concurrent processes.