Use the htsky library in place of htrdr_sky.<c|h>
Add the dependency to the HTSky library and Remove the now useless
dependency to the HTGOP, HTCP and HTMie libraries that are no more used
directly. Refactor the code to use the HTSky library rather than
htrdr_sky. Finally remove the support of the "cache grids" option that
is not supported by the HTSky library.