Remove a FATAL in the CIE XYZ sampling function:
The function samples a wavelength by first selecting a sub-band into the
CIE XYZ spectral range and then by continuously sampling the CIE
tristimulus curve into this sub-band. Due to numerical imprecisions,
this latter computation can find a wavelength that is actually not
included into the sampled sub-band, leading to an error that previously
was handled by a FATAL.

This commit replaces it by a warning message and a fallback procedure
that in such unforeseen situation selects the wavelength and the center
of the sampled sub-band.