Ray filter function for the combustion chamber geometry
Implement 2 ray filter functions that ignore the first or last
intersected surface. The first function will be used to allow the rays
of the camera to enter the combustion chamber while the second will be
used to reach the laser surface emission during the backward ray tracing,
i.e. to let the laser sheet enter the combustion chamber.