Add the thin-lens camera model
Add the lens-radius, focal-dst, and focal-length options to the '-C'
option that defines a camera. By default the camera is a pinhole until
the 'lens-radius' is set to a value greater than 0. With a thin-lens
camera, the 'focal-dst' control the distance to focus and 'focal-length'
is another way to define its field of view.