htrdr-planeto: merge sw and lw radiance calculations
In this commit, the shortwave algorithm is updated to no longer track
absorption transmissivity. As in longwave, an extinction distance is
sampled and then it is defined whether this distance corresponds to
diffusion or absorption. With this update, we can easily merge shortwave
and longwave algorithms in the same function. Additionally, we avoid
tracing one ray per scatter event and thus speed up shortwave path
tracing.