htrdr-planeto : upd the definition of spectral integration
This commit removes the reference temperature from the option that
defines a Planck spectral integration. In shortwave, we now use the
temperature of the source as the reference temperature of the Planck
function. In longwave, the reference temperature is set to the maximum
ground temperature which should be the maximum temperature of the
system.