htrdr-planeto: source radiance by wavelength
Source radiance can now be described by wavelength by providing a new
file following a file format described in this commit. When set,
this new argument overrides the source temperature.

In any case, these data are loaded, but are not used internally either
to sample a wavelength during rendering, nor to return spectrally
varying radiance when querying the source. Those are the next steps.