Update error handling in htrdr_ran_wlen_planck_create
In an error occurs, the function would simply free the created structure
without assigning NULL to the returned pointer. As a result, the caller
could think that the structure was still allocated and try to free it,
leading to a double freeing of memory and thus a crash. In this commit,
we correct this problem by simply returning a NULL pointer when an error
has been detected by the function.