atmosphere/planeto: upd of the planck/CIE cumulative building
To speed up the sampling of the spectral dimension, we divide the Planck
function and the CIE tristimuli into small bands of fixed size, about 1
nanometre in length. We also require that the spectral bands of the
underlying semi-transparent material include at least one of these
sampling bands. This latter constraint becomes problematic, however,
when the integration domain begins or ends near the end or beginning of
a spectral band: the spectral bands truncated at the integration domain
can be very thin, resulting in a prohibitive total number of sample
bands.

We are unable to find the justification for the requirement to have at
least one sample band in a spectral band. Indeed, the two are distinct
since the sample bands are used to sample a wavelength which is then,
and only then, used to recover the spectral band of the semi-transparent
material in which it is located, before finally sampling a quadrature
point in that spectral band. In this commit, we therefore remove this
constraint to avoid having too many sample bands, which could saturate
the memory space and cause an allocation error.