Write a POSIX Makefile to replace CMake
The build procedure is written in POSIX make, which the user can
configure via the config.mk file. The make.sh script contains commands
that could be found directly in POSIX make, but which are placed here to
simplify writing the Makefile.

pkg-config is used for internal and external dependencies. For internal
dependencies, we write the pc file of the local libraries, even if they
are only used by the htrdr build system. In fact, this is a pleasant and
practical way of managing their recursive dependencies, as required by
static libraries.

In addition to the features already provided in its CMake alternative,
this Makefile supports the use of static libraries and provides an
uninstall target. In any case, the main motivation behind its writing is
to use a good old well-established standard with simple features,
available on all UNIX systems, thus simplifying its portability and
support while being much lighter