Translate htrdr-combustion man page to mdoc
In the course of translation, its content has been updated. In
particular, we have added the missing "EXIT STATUS" section. References
to the MPI 2.0 and OpenMP 2.0 specifications have been added to the new
STANDARDS section. The COPYRIGHT and LICENSE sections have been removed,
as they are not part of conventional Linux manuals (mandoc doesn't
mention them at all). Finally, the entire text has been revised to make
it clearer and more in line with mandoc conventions.