Partially link static library object files
Make hidden symbols local to the relocatable object to avoid conflicts
with internal symbols in another static library. In release mode, delete
all symbols not required for relocation processing, but keep them in
debug mode for.

Replace the HTRDR_CORE_API macro with the HTRDR_API macro. The former
was used by the core library, while the latter is intended for internal
libraries. The core library can be compiled as a shared or static
library, while internal libraries are static only. But in fact, both
static and shared libraries can be compiled using the same compilation
flags and definitions. In fact, this is what is done so that hidden
symbols are local to the static library. The HTRDR_CORE_API macro is
unnecessary and is therefore removed.