Update compiler and linker flag definitions
In both cases, there are two sets of macros. <C|LD>FLAGS_SO (for shared
objects) and <C|LD>FLAGS_EXE (for executables) defined from the
<C|LD>FLAGS_DEBUG or <C|LD>FLAGS_RELEASE macro. This commit (re)adds the
<C|LD>FLAGS macros that are by default defined as <C|LD>FLAGS_DEBUG or
<C|LD>FLAGS_RELEASE, depending on the value of BUILD_TYPE. Macros
<C|LD>FLAGS_SO and <C|LD>FLAGS_EXE are now defined from them.

This allows the compiler and linker flags for executables and shared
objects to be overwritten simply by updating the CFLAGS and LDFLAGS
macros. What's more, this is in line with the expected way of
proceeding, CFLAGS and LDFLAGS being standard macros.