planeto: update edition of the examples in the manual
Ensure that command lines do not exceed 78 columns, which is mandoc's
default text width for ASCII output. This prevents incorrect formatting,
i.e. lines exceeding the page width, when text is output in other
formats such as pdf or html.

Command lines are split in a specific way. We have therefore added a
paragraph at the start of the examples section to highlight the presence
or absence of spacing in the command line breakdown. The aim is to help
the user understand the error of an approximate copy and paste
operation, adding spaces where none are needed and rendering the command
line invalid.