Improve Makefile readability and portability
Group all pc dependencies for each executable into a single call to the
pkg-config utility. This not only reduces the (small) workload
associated with multiple calls to pkg-config, but also makes the way
CFLAGS and LIBS dependencies are defined more concise.

Simplify prerequisite checking by displaying only the pkg-config command
line used to check the version rather than displaying a handwritten
error message if it fails. If one of the commands fails, the
configuration step stops at the given command, which displays the
corresponding error message. There is therefore no need to do anything
further.

Addition of macros enabling the user to define the subpaths in which
binaries, libraries, documentation and manuals are installed, in
relation to the PREFIX macro. Depend ending on the system, the expected
file hiearchy can be different.

Remove distclean targets. These were not only misused, as they deleted
files that should have been cleaned up by the normal "clean" target. But
they were also useless because no dist target is provided. And there are
no plans to provide one, since htrdr is designed to be installed from
source.