core: fix MPI lock when gathering image
If the number of tiles to render was not a multiple of the number of
processes performing the calculation, the remaining tiles were poorly
managed: they were all supposed to be calculated by the master process
and, more importantly, some of them were wrongly rejected. This commit
not only corrects the latter problem, but also makes a more smart
distribution of the remaining tiles by dividing them between all the
processes.