planets: add and document the -r option
It is used to calculate the volumic radiative budget, not an image.
Currently, only the analysis of the option is performed. Its behavior is
not yet implemented, as mentioned in the error message.

The manual page has been updated accordingly. Not only to document this
new option (and its result) but also to correct the fact that
htrdr-planets is no longer just a rendering engine.

When modifying the man page, its formatting was slightly updated to
improve the readability of its sources. In addition, the use of the Pa
macro has been corrected: whereas it should only be used for paths, it
is used for arguments (whose value is a path).