core: refactoring the distribution of MPI work
Prepare to add a solve function that will not distribute rendering but
more generic calculations.

The distribution of MPI work thus becomes an internal API of the core
library. What's more, it no longer manages tiles, but chunks, a more
agnostic term. The chunk is identified by a 64-bit unsigned integer,
whereas previously a tile was encoded on 32 bits. Again, this is to make
the system more generic: tiles were a set of pixels in an image, whereas
chunks can be anything (e.g. a set of realisations), and their number
could therefore exceed a 32-bit encoding.