planets: fixed dump octree option when used with longwave
htrdr returned an error due to the configuration of the spectral
dimension which, in longwave, queried the ground temperature, a ground
which was not loaded when the dump octree option was activated. However,
as with the ground, there's no need to set the spectral dimension when
the dump octree option is active.