planets: update the manual of the -V option
Update of the -V option to configure how acceleration structures are
built. Currently, only the manual is updated. The code remains
unchanged. In other words, the documentation is ahead of the expected
behaviour of the code.

This option must also allow to control the processes that build the
acceleration structures. When using MPI, the common case on a
supercomputer is to use shared disk space between processes. So, when
storing acceleration structures on disk, all processes build their own
set of acceleration structures and save them in the same file, since
this space is the same for all processes. And although no errors were
reported by users, this was only by chance since, depending on MPI
scheduling, a process could open a non-empty storage space. In such a
situation, it would have expected a valid storage space and would have
returned an error because the octrees were actually under construction.
In the best-case scenario, there were no errors reported, and all the
processes were writing the same data to the same file.  This
considerably increased pressure on the disk and reduced construction
performance. In fact, in such a situation, the user would not have had
to store the accelerator structures to force all the processes to build
their own set of structures in memory.

To avoid these problems, the acceleration structure will now only be
built by the master process, assuming that the same disk space is shared
by all processes. If this is not the case, the user must explicitly
activate the build for all processes.