planets: configure the atmosphere and multi-proc
Update to the way atmopshere is configured when using MPI. In fact,
manage the new way of configuring the acceleration structure as
documented in commit 0103104.

The analysis of the input arguments is therefore updated to take account
of these. And, by default, if octree storage is used, octrees are now
built by the master process only, assuming that storage is on a disk
shared by the different processes.