Rewriting the Makefile
Divide the Makefile into sub-makefiles, one for each command, and an
additional one for the base library. This makes the build system easier
to read and manage. It also simplifies the way in which commands are
conditionally built: now it's just a matter of executing the command's
makefile according to the state of the macro that controls its
building.

Finally, the htrdr executable has been rewritten as a POSIX shell. The
use of C was perfectly dispensable here, given its sole purpose: to run
the corresponding command.