core: Fix of the "uninstall" target in the Makefile
A quotation mark was not closed.

The Makefile.core file is executed like the others rather than being
explicitly invoked in the common targets of the main Makefile. The only
reason for not doing so was that its default target was invoked by the
other Makefiles. Now, all its other targets must be explicitly invoked
(clean, install, uninstall). Hence this change