core: clean static and shared libraries
The Makefile clean target now deletes static and shared versions of the
core library, regardless of the value of the LIB_TYPE macro. This
prevents an old shared library from being used instead of the static
library that has been defined as the one to be built and used. In fact,
the pkg-config utility would prefer its shared version, even if the
--static option is used.