Make the commands a simple wrapper for their library
The commands were compiled in all cases. But their behaviour was defined
by the compilation process depending on whether the commands were
enabled. Now, when disabled, commands are simply not compiled.  The user
will therefore encounter an error when using them, which the system will
notify by indicating that the command is not present. Previously, it was
the command itself that returned an error.

Even if this "disabled" executable gave an explanation as to why it
could not be used, such an implementation was unfortunate given the
intended functionality, which was simply to notify the user that they
were unable to use the command. Which the system already does very well
without further development.