Decomposition of the radiative budget
The realization function for the radiative budget has been modified in
such a way that its direct and diffuse components have been
differenciated. The direct component of the radiative budget only has to
sample the solar cone, while the diffuse component still samples the
unit sphere. This will reduce the variance associated to the total in
the case the external source is restricted to a very small solid angle.